:recycle: Extract report domain to module This extraction will be useful for creation of monitoring module, which will only depend on this domain module
