+++ modules/core/src/main/kotlin/datamaintain/core/step/check/rules/contracts/ScriptCheckRule.kt
@@ -1,8 +1,8 @@
 package datamaintain.core.step.check.rules.contracts
 
 import datamaintain.core.exception.DatamaintainCheckException
-import datamaintain.core.script.ScriptWithContent
-import datamaintain.core.step.check.rules.CheckRule
+import datamaintain.domain.CheckRule
+import datamaintain.domain.script.ScriptWithContent
 
 abstract class ScriptCheckRule: CheckRule {
     /**
