+++ modules/core/src/main/kotlin/datamaintain/core/step/check/rules/contracts/ScriptWithContextCheckRule.kt
@@ -1,9 +1,9 @@
 package datamaintain.core.step.check.rules.contracts
 
 import datamaintain.core.exception.DatamaintainCheckException
-import datamaintain.core.script.ExecutedScript
-import datamaintain.core.script.ScriptWithContent
-import datamaintain.core.step.check.rules.CheckRule
+import datamaintain.domain.CheckRule
+import datamaintain.domain.script.ExecutedScript
+import datamaintain.domain.script.ScriptWithContent
 
 abstract class ScriptWithContextCheckRule(
         executedScripts: Sequence<ExecutedScript>
