+++ modules/core/src/main/kotlin/datamaintain/core/step/check/rules/implementations/AlwaysFailedCheck.kt
@@ -1,9 +1,9 @@
 package datamaintain.core.step.check.rules.implementations
 
 import datamaintain.core.exception.DatamaintainCheckException
-import datamaintain.core.script.ScriptWithContent
-import datamaintain.core.step.check.rules.ScriptType
 import datamaintain.core.step.check.rules.contracts.ScriptCheckRule
+import datamaintain.domain.ScriptType
+import datamaintain.domain.script.ScriptWithContent
 
 class AlwaysFailedCheck: ScriptCheckRule() {
     override fun check(script: ScriptWithContent) {
