+++ modules/core/src/main/kotlin/datamaintain/core/step/executor/Execution.kt
@@ -1,5 +1,5 @@
 package datamaintain.core.step.executor
 
-import datamaintain.core.script.ExecutionStatus
+import datamaintain.domain.script.ExecutionStatus
 
 data class Execution (val executionStatus: ExecutionStatus, val executionOutput: String? = null)
