+++ modules/core/src/main/kotlin/datamaintain/core/step/sort/SortingStrategy.kt
@@ -1,6 +1,6 @@
 package datamaintain.core.step.sort
 
-import datamaintain.core.script.Script
+import datamaintain.domain.script.Script
 
 abstract class SortingStrategy<U>() {
     abstract fun <T : Script> sort(scripts: List<T>, getter: (T) -> U): List<T>
