+++ modules/core/src/test/kotlin/datamaintain/FilterTest.kt
@@ -8,10 +8,10 @@ import datamaintain.core.config.DatamaintainScannerConfig
 import datamaintain.core.db.driver.DatamaintainDriver
 import datamaintain.core.db.driver.FakeDriverConfig
 import datamaintain.core.script.FileScript
-import datamaintain.core.script.ScriptAction
-import datamaintain.core.script.Tag
 import datamaintain.core.step.Filter
 import datamaintain.core.step.executor.ExecutionMode
+import datamaintain.domain.script.ScriptAction
+import datamaintain.domain.script.Tag
 import io.mockk.mockk
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
