+++ modules/core/src/test/kotlin/datamaintain/core/db/driver/FakeDatamaintainDriver.kt
@@ -1,8 +1,8 @@
 package datamaintain.core.db.driver
 
-import datamaintain.core.script.ExecutedScript
-import datamaintain.core.script.ScriptWithContent
 import datamaintain.core.step.executor.Execution
+import datamaintain.domain.script.ExecutedScript
+import datamaintain.domain.script.ScriptWithContent
 
 
 class FakeDatamaintainDriver : DatamaintainDriver("") {
