+++ modules/core/src/test/kotlin/datamaintain/core/script/TagMatcherTest.kt
@@ -1,5 +1,6 @@
 package datamaintain.core.script
 
+import datamaintain.domain.script.Tag
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import strikt.api.expectCatching
