+++ modules/domain-report/src/main/kotlin/datamaintain/domain/script/ExecutionStatus.kt
@@ -1,4 +1,4 @@
-package datamaintain.core.script
+package datamaintain.domain.script
 
 enum class ExecutionStatus {
     OK,
